Swapped header and source files' contents My bad
