+++ src/euclidean.c
@@ -1,8 +1,13 @@
-#ifndef DISTANCE_H
-#define DISTANCE_H
+#include "distance.h"
+#include <math.h>
 
-#include <stddef.h>
+double distance(const double *p, const double *q, size_t size) {
+    double sum_of_squares = 0.0;
 
-double distance(const double *p, const double *q, size_t size);
+    for (size_t i = 0; i < size; ++i) {
+        double diff = p[i] - q[i];
+        sum_of_squares += diff * diff;
+    }
 
-#endif
+    return sqrt(sum_of_squares);
+}
